Set layout for content section
Used nesting grids to design and layout the main content section as well
as the right "sidebar" with the trending and announcement sections.
Everything is still filled in with dummy text and more needs to be
styled with CSS which I will do next, but now the layout for the
dashboard is pretty solid.